Problems on report controller, foreing keys

+++ server/models/report.ts
@@ -1,19 +1,19 @@
-import { Model, DataTypes } from "sequelize";
-import { sequelize } from "./conn";
+import { Model, DataTypes } from 'sequelize';
+import { sequelize } from './conn';
 
 export class Report extends Model {}
 Report.init(
-  {
-    // Model attributes are defined here
-    fullname: {
-      type: DataTypes.STRING,
-    },
-    phone_number: {
-      type: DataTypes.INTEGER,
-    },
-    report_data: {
-      type: DataTypes.STRING,
-    },
-  },
-  { sequelize, modelName: "report" }
+	{
+		// Model attributes are defined here
+		fullname: {
+			type: DataTypes.STRING,
+		},
+		phone_number: {
+			type: DataTypes.BIGINT,
+		},
+		report_data: {
+			type: DataTypes.STRING,
+		},
+	},
+	{ sequelize, modelName: 'report' },
 );
